+++ implementations/java/compression/src/utils/fractal/mapping/FractalMapping.java
@@ -1,0 +1,5 @@
+package implementations.java.compression.src.utils.fractal.mapping;
+
+public record FractalMapping(int rangeX, int rangeY, int domainX, int domainY, float s, float o) {
+
+}